requirements.txt
flask
flask-login
flask-wtf
flask-sqlalchemy
flask-bcrypt
email-validator
